backend/requirements.txt
Flask
google-genai
Pillow
Flask-CORS